slackbot/requirements.txt
slackclient
SQLAlchemy
psycopg2-binary